Update CMS_project_v1_config.py
Changed the precision (from 16-mixed to 32-true) in the CMS config.

+++ workspaces/CMS_workspace/CMS_project_v1/config/CMS_project_v1_config.py
@@ -14,7 +14,7 @@ def set_config(c):
     c.lr_scheduler = True
     c.save_error_bounded_deltas = False
     c.error_bounded_requirement = 10
-    c.precision = "16-mixed"
+    c.precision = "32-true"
 
     # === Additional configuration options ===
